+++ packages/marketing-modal/src/index.ts
@@ -1,2 +1,3 @@
-export { default, GraphicStyle, BlobPosition } from './MarketingModal';
+export { default } from './MarketingModal';
+export { GraphicStyle, BlobPosition } from './MarketingModal.types'
 export { default as svgBlobs } from './blobs';
